requirements.txt
cookiecutter
oslo.utils
pycrypto
ruamel.yaml
